cambiato in float colonna salary

+++ database/migrations/2019_10_01_142625_create_employees_table.php
@@ -20,7 +20,7 @@ class CreateEmployeesTable extends Migration
             $table->char('lastname');
             $table->date('birthDay');
             $table->char('jobTitle');
-            $table->integer('salary');
+            $table->float('salary');
         });
     }
